Split input related estimator tests into a new file to fix timeout issue. Change: 146735638
